Moved method lower in the class heirarchy
