+++ Tools/SharePoint.Modernization/SharePointPnP.Modernization.Framework/Transform/BaseTransform.cs
@@ -1,7 +1,9 @@
-﻿using SharePointPnP.Modernization.Framework.Telemetry;
+﻿using Microsoft.SharePoint.Client;
+using SharePointPnP.Modernization.Framework.Telemetry;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
+using System.Net;
 
 namespace SharePointPnP.Modernization.Framework.Transform
 {
@@ -169,5 +171,82 @@ namespace SharePointPnP.Modernization.Framework.Transform
                 observer.SetPageId(pageId);
             }
         }
+
+        #region Helper methods
+
+        /// <summary>
+        /// Gets the version of SharePoint
+        /// </summary>
+        /// <param name="clientContext"></param>
+        /// <returns></returns>
+        internal SPVersion GetVersion(ClientRuntimeContext clientContext)
+        {
+            try
+            {
+                Uri urlUri = new Uri(clientContext.Url);
+                HttpWebRequest request = (HttpWebRequest)WebRequest.Create($"{urlUri.Scheme}://{urlUri.DnsSafeHost}:{urlUri.Port}/_vti_pvt/service.cnf");
+                request.UseDefaultCredentials = true;
+
+                var response = request.GetResponse();
+
+                using (var dataStream = response.GetResponseStream())
+                {
+                    // Open the stream using a StreamReader for easy access.
+                    using (System.IO.StreamReader reader = new System.IO.StreamReader(dataStream))
+                    {
+                        // Read the content.Will be in this format
+                        // SPO:
+                        //vti_encoding: SR | utf8 - nl
+                        //vti_extenderversion: SR | 16.0.0.8929
+                        // SP2019:
+                        // vti_encoding:SR|utf8-nl
+                        // vti_extenderversion:SR|16.0.0.10340
+                        // SP2016:
+                        // vti_encoding: SR | utf8 - nl
+                        // vti_extenderversion: SR | 16.0.0.4732
+                        // SP2013:
+                        // vti_encoding:SR|utf8-nl
+                        // vti_extenderversion: SR | 15.0.0.4505
+                        // Version numbers from https://buildnumbers.wordpress.com/sharepoint/
+
+                        string version = reader.ReadToEnd().Split('|')[2].Trim();
+
+                        if (Version.TryParse(version, out Version v))
+                        {
+                            if (v.Major == 14)
+                            {
+                                return SPVersion.SP2010;
+                            }
+                            else if (v.Major == 15)
+                            {
+                                return SPVersion.SP2013;
+                            }
+                            else if (v.Major == 16)
+                            {
+                                if (v.MinorRevision < 6000)
+                                {
+                                    return SPVersion.SP2016;
+                                }
+                                else if (v.MinorRevision > 10300)
+                                {
+                                    return SPVersion.SP2019;
+                                }
+                                else
+                                {
+                                    return SPVersion.SPO;
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+            catch (WebException ex)
+            {
+                // todo
+            }
+
+            return SPVersion.SPO;
+        }
+        #endregion
     }
 }
